backend/requirements.txt
django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
psycopg2-binary==2.9.11
python-dotenv==1.2.2
django-cors-headers==4.9.0
